requirements.txt
torch>=2.1
torchvision>=0.16
PyYAML>=6.0
numpy>=1.24
pandas>=2.0
tqdm>=4.66
